requirements.txt
PyYAML

numpy
matplotlib
PIL
tqdm

torch
torchvision